Removed address and port from prod instanceserver websocket request
As part of IR-5570, the address and port of the instanceserver that a websocket
connection is being made to are no longer sourced from query params on the WS
address. Instead, the nginx server now looks up the instance record of the
already-provided instanceID and proxies to the address in the record, if the
record exists.

As the query params are no longer used, this removes them from the connection.
In developments settings, they are still used as the direct IP:port of the
connection since there normally is no nginx server to act as an intermediary.