Merging prd hotfixes into stg (#905)
* RC0.9 Ir 5328 users don t spawn on spawn point in published scenes seem to spawn at origin (#602)

* Add transform component as a dependency of spawning avatar to ensure pose is correct

* fix import

---------

Authored-by: HexaField <joshfield999@gmail.com>

* Set max location to 5 (#612)

* Set max location to 5

* message

* Release0.9 Hotfix Env Maps (#686)

* HOTFIX0.9 Add envmap fixes into release0.9 (#683)

* Add envmap fixes into release0.9

* GLTF transform dependencies in common

---------

Co-authored-by: Michael Estes <michael@etherealengine.com>

* R0.9 Envmap editor hotfix (#685)

Update envmap editor with fix from dev

---------

Co-authored-by: Michael Estes <michael@etherealengine.com>

* hotfix to pin helmet version (#689)

* Additional hotfixes from release0.9 to prd (#716)

* HOTFIX0.9 Add envmap fixes into release0.9 (#683)

* Add envmap fixes into release0.9

* GLTF transform dependencies in common

---------

Co-authored-by: Michael Estes <michael@etherealengine.com>

* R0.9 Envmap editor hotfix (#685)

Update envmap editor with fix from dev

* ServerLogger.ts mv all endpoints behind endvars; remove net check (#711)

* Update package.json

---------

Co-authored-by: Daniel Belmes <3631206+DanielBelmes@users.noreply.github.com>
Co-authored-by: Michael Estes <michael@etherealengine.com>
Co-authored-by: infinitelos <150717186+infinitelos@users.noreply.github.com>

* Hotfix 12/17/24 (#850)

* HOTFIX0.9 Add envmap fixes into release0.9 (#683)

* Add envmap fixes into release0.9

* GLTF transform dependencies in common

---------

Co-authored-by: Michael Estes <michael@etherealengine.com>

* R0.9 Envmap editor hotfix (#685)

Update envmap editor with fix from dev

* ServerLogger.ts mv all endpoints behind endvars; remove net check (#711)

* Update package.json

* Release0.9 pin react icons (#778)

Pin react-icons to 5.0.1

* Fixed some security issues with identity-provider service (#807)

Create method was able to set fields on new guests that it shouldn't.
It now only preserves id, token, type, userId, createdAt, and updatedAt.
Also made sure that guest tokens are UUIDs, as one was able to put any
value for guest tokens. Now, if that value isn't a UUID, it is replaced
with a UUID.

Disabled external patching of identity-providers. This was not being
done externally, but it was allowing patching of fields that shouldn't
be patched. In general, this is not something that users should be doing,
hence the disabling.

Added some identity-provider tests and organized them by method.

Resolves IR-5581

* Revert "Fixed some security issues with identity-provider service" (#832)

Revert "Fixed some security issues with identity-provider service (#807)"

This reverts commit 24ff7cd4181901ff4128d4c2e1658dee12857974.

* Fixed some security issues with identity-provider service (#833)

Create method was able to set fields on new guests that it shouldn't.
It now only preserves id, token, type, userId, createdAt, and updatedAt.
Also made sure that guest tokens are UUIDs, as one was able to put any
value for guest tokens. Now, if that value isn't a UUID, it is replaced
with a UUID.

Disabled external patching of identity-providers. This was not being
done externally, but it was allowing patching of fields that shouldn't
be patched. In general, this is not something that users should be doing,
hence the disabling.

Added some identity-provider tests and organized them by method.

Resolves IR-5581

* Ir 5663 mobile warnings for release0.9 (#777)

* warning alert for unsupported browser on locations (#443)

* warning alert for unsupported browser on loations

* usebrowsercheck hook

* add hook file

* add .value

* remove unused code

* fix import

---------

Co-authored-by: Daniel Belmes <danielbelmes@ymail.com>

* [IR-5684] ] warning message on location for mobile devices (#726)

* warning message on mobile devices

* update alert copy

* IR-5663 fix import

---------

Co-authored-by: Andy Chen <44120813+achen5671@users.noreply.github.com>
Co-authored-by: Kyle Baran <kbaran@bitscoop.com>

* Hardened security around file-browser service (#763)

IR-5564 IR-5568 Hardened security around file-browser service (#722)

Hardened security around file-browser service

File-browser services were letting users use relative paths to affect files
anywhere in the directory structure, including projects they did not have
access to, or files outside of the project structure altogether.

File-browser services now resolve the path being accessed, including any
relative parts of the path, and then check if a) that's within a project's
assets or public folder and b) the user has owner or editor permissions on
that project.

Added checks that directories, file names, and file extensions do not have
invalid characters, and that file names and extensions fall within specific
lengths.

Resolves IR-5564 and IR-5568

---------

Co-authored-by: Daniel Belmes <3631206+DanielBelmes@users.noreply.github.com>
Co-authored-by: Michael Estes <michael@etherealengine.com>
Co-authored-by: infinitelos <150717186+infinitelos@users.noreply.github.com>
Co-authored-by: Andy Chen <44120813+achen5671@users.noreply.github.com>

---------

Co-authored-by: Daniel Belmes <3631206+DanielBelmes@users.noreply.github.com>
Co-authored-by: Michael Estes <michael@etherealengine.com>
Co-authored-by: infinitelos <150717186+infinitelos@users.noreply.github.com>
Co-authored-by: Andy Chen <44120813+achen5671@users.noreply.github.com>