Removed auto-deletion of login-tokens (#1453)
Some email forwarding services, and seemingly Apple devices, will follow
links that they detect, which would use up the token and cause valid
attempts to follow the link to fail because the token was already used.

Set expiration of magiclink login tokens to 10 minutes.

Added a line to email template to note that the links are only valid for 10 minutes.