Fixed bug in getServerPodInfo and getServerContainerInfo (#1494)
getServerPodInfo was passing containerStatuses of server pod assuming
it was defined, but failed/evicted pods have no containerStatuses.
This was causing getServerContainerInfo to error out when attempting
to call .map on undefined.

Made getServerPodInfo default to an empty array if containerStatuses
is undefined/null. Also added a conditional check on
getServerContainerInfo's .map so that it'll just return null if somehow
no items are passed to it.