[IR-8732] Unit Tests - GLTF Loading: Asset (#1684)
* new: GLTFValidate helpers (sketch)

@note Current implementation completely ignores this root property.
But it is mandatory by spec to exist, have a version and validate it.
This PR is a preparation for fixing that problem.

* tst: Unit Tests for `GLTF.loader.asset`

* tst: Fix missing import typo mistake

* tst: Add valid cases that should not throw for optional Asset props

* fix: Attempt to fix regex vulnerability issue

* fix: Change `glTF.asset.version` regex to use bounded quantifiers

Removes the chance of ReDos attacks happening due to unbounded backtracking.
Bounded quantifiers are denoted by {min,max} syntax,
where min and max are integers that specify the minimum and maximum number of repetitions.
https://plainenglish.io/blog/preventing-regular-expression-backtracking-in-javascript

* fix: Typo in bounded quantifiers syntax

* fix: Reduce permutation count in the asset.version bounded quantifier